Allow size 4 for the program buffer
Why is Program Buffer only allowed to be size 8?

In our previous implementation, for an XLEN=32 design, we only needed 7 words to implement the desired functionality for both program buffer and data transfer. With the latest "quick access" proposal we'll need 3 words for data. Is it possible to do what we need with only 4 words? (or 5 if we must implement ebreak at the end)?