Introduce resumereq command, similar to haltreq.
This is needed to deal with the following race:
A hart is running, and the debugger wants to select a different hart, by
writing \Fhartsel. In the old design the debugger would have to write
\Fhaltreq low, because the hart is running. But what if the hart halts
just before the write happens? Now the debugger needs to write \Fhaltreq
high. It cannot know in advance what to write.

In the new scheme it writes \Fhaltreq and \Fresumereq low whenever it
doesn't want anything to happen.