Fix ambiguity in busy error reporting.
The problem is as follows:
1. Debugger writes to sbdata0, triggering a write.
2. Debugger writes to sbdata0 while the access is still in progress. A
busy error needs to be set.
3. The original access finishes with eg. an alignment error.

Before this change, we would have to pick which error (alignment/busy)
to report, while for full understanding the debugger needs to know both.
This is fixed by adding a sbbusyerror bit, which is used to indicate
busy errors, leaving sberror for other error reporting.