Add support for 2^20 harts, up from 2^10
The main difference is to remove the halt region in favor of 4
registers. This is slightly slower to determine which hart is halted if
their number is between 32 and 1024, but it allows for many more harts
overall, and the time to figure out which one is halted is O(log(N)).