Change dcsr.prv reset value to 3
This will help avoid a trap where an implementation might end up with
prv 0 when entering debug out of reset. Even in that case prv should be
set from the privilege mode of the processor, which should be machine
mode, but specifying this as the actual value makes things a little
simpler.