Make haltreq and resumereq proper write-only.
They already mostly were, but then I clarified some stuff and broke
keeping haltreq high during reset, which is the only mechanism to catch
a hart out of reset if optional reselthaltreq isn't implemented.

Also:
* resume ack is cleared by the DM when resumereq is written 1.
* haltreq can be high together with hartreset.

This is the minimal version of #378. It has all the same functionality,
but lacks some of #378's clarifications, which may be making this harder
to review.