Stopcount only applies to hart-local counters.
The cycle counter is global across all harts in a core, so shouldn't
stop in multi-hart cores.

This change is not 100% backwards compatible. A testsuite testing 0.13
might check that cycle stops while halted with stopcount=1, and this
change would break that.
I think it's compatible enough, however, since halting cycle on a
multi-hart core with not all harts halted is even more broken.