Add nmi bit to etrigger. (#408)
* Add nmi bit to etrigger.

Fixes #407.

* nmi ignores m/s/u.

The rationale is that nobody can think of a reason that somebody might
want to only trigger on NMI in a specific mode, and by ignoring those
bits it's possible to use a single trigger to trigger on e.g. illegal
instructions in U mode, but on all NMIs.