Abstract command clarifications.
In Section 3.7 Abstract Commands:
Clarify cmderr=2 (not supported) if not supported at all,  vs.
cmderr=4 (halt/resume) if supported but not in the current hart state.

Then in subsections for each abstract command, clarify that their
sequence of operation is conditional on their being supported as
described above.  This makes it clear for example that cmderr=2 is
checked before other error conditions.

"may not" can mean "must not", reword to avoid ambiguity.

Fix section references in cmdtype table.