Tighten up trigger specification. (#478)
* Tighten up trigger specification.

This contains a bunch of improvements that came out of an e-mail
discussion. The biggest one is that it specifies the priority of various
triggers compared to the exceptions already in the Privileged Spec.

* dmcontrol->mcontrol, fix cut and paste mistakes.

* All trigger fields are WARL.

It already said this in Trigger Registers: "All {\tt tdata} registers
follow write-any-read-legal semantics."

* supported->legal

Co-Authored-By: Paul Donahue <48959409+pdonahue-ventana@users.noreply.github.com>

* Incorporate minor feedbacks.

* Trigger PC is set to next instruction to execute.

Previously on precise timing=after exceptions we said it should be set
to the PC of the matching instruction. But that doesn't work well
because the debugger can't know if triggers are precise or not.

* Tweak trigger priorities.

Add itrigger/etrigger.
Move mcontrol after (on previous instruction) to top priority.

* Fix typo.

Co-Authored-By: Paul Donahue <48959409+pdonahue-ventana@users.noreply.github.com>

* Clarify mcontrol load before side effect.

* Clarify trigger priority table.