Make haltsum0 optional if there is only one hart. (#505)
* Make haltsum0 optional if there is only one hart.

This is not forwards compatible. Debuggers that rely on haltsum0 would
break with this change. But it makes the debug interface on tiny 1-hart
systems smaller, especially because it allows the number of address bits
in the DMI to be 1 less. (Haltsum0 at 0x40 due to other efforts to
maintain backwards compatibility.)

Fixes #389.

* Make language more precise.