Debuggers should know when harts are unavailable. (#520)
* Debuggers should know when harts are unavailable.

Make unavail sticky, needing to be acknowledged with new ackunavail bit.

This is a backwards incompatible change. I think it's OK because
implementations that support unavailable harts are rare (I don't know of
any), and the change only means that debuggers must be updated. If
hardware doesn't implement this then everything will still work,
although users might not be notified when harts become unavailable.

Fixes #461.

* Add stickyunavail bit.

Every other way to ensure hardware backwards compatibility was getting
too convoluted.

* Clean up a little.

Keep definition of the bits in the register definition, and definition
of states in the DM section.