Make things less ambiguous.
Remove explanation about how icount can be done in a single bit.

The single-bit implementation can work, but only if you tie all mode
bits together and have max count of 1. If all mode bits are enabled
together, there is no case where the existence of a pending bit can make
a different. But this is only useful for an external debugger, and we
already say that for external debuggers dcsr.step is the preferred
solution.

For native debugging, the least hardware I can think of is 2 bits,
because there are effectively 3 states: disabled, pending, and enabled.
This assumes the mode bits are tied together in such a way that there is
only one sensible option. I'm not sure if that's even worth mentioning
anywhere. It would be a long explanation with little benefit.