Make data registers Message Registers.
Message Registers allow data to be sent to the other side, but not to be
read back by the side that wrote the data.

This allows smaller FPGA implementations by duplicating the storage and
removing muxes in the design.

It is incompatible with old debuggers, if they use these registers in
surprising ways.