Add trigger action values, for debug_defines.h.
Leave the descriptions empty, and change registers.py to not emit
anything for values whose descriptions are empty.

I'm not a huge fan of the duplication of those values in the XML file,
but I don't want to invent a new mechanism to avoid it right now. (Ooh,
we could use M4 to generate the XML and use an existing mechanism...)