Remove license from debug_defines.h.
CC BY 4.0 is incompatible with the GPL, meaning the file could not be
used in GPL projects like OpenOCD. RISC-V has plans to address this, but
the lawyers are working on other issues right now. Until that is
properly resolved, we're reverting to the way things were before #641
and removing the license line.