Add mcontrol6.{uncertain,uncertainen} bits.
Uncertain trigger matches occur when the hardware can't do a precise
comparison, e.g. on batched vector accesses, and far AMO data.

This lets the user select whether they want the trigger to fire if a
match is uncertain, and the hardware can indicate whether a given match
was certain or not.

Lots of mailing list discussion prompted this:
https://lists.riscv.org/g/tech-debug/topic/94864865#1030