Remove timing bit again. Introduce tinfo.version.
Lots of email discussion, subject:"removed mcontrol6.timing,
compatibility"

Now that hit0/hit1 clearly describe when a trigger hit, we don't need a
separate mechanism to communicate whether triggers hit before/after. The
only remaining benefit timing provided was that it would let the
debugger know when the trigger is set what the timing behavior would
probably be. (Emphasis on probably, since it's not possible to honor the
requested timing in all situations.)