Sdtrig: Clarify that tinfo.version is a global constant, not per-trigger
The purpose of "tinfo.version" field is apparently to describe the version
of the whole Sdtrig implementation. That is, it is a constant unaffected by
the current value of tselect CSR. Judging form the original commit that
introduced it (0374fd6d9b3e8675694bf99d613f455bee264fae) and from the
related mailing list discussion
(https://lists.riscv.org/g/tech-debug/topic/96888390).

However, the sentences

`This register provides access to the trigger
selected by tselect. The reset values listed here apply to
every underlying trigger.`

leave certain room for interpretation that "tinfo.version" is
a per-trigger constant, not a global one, and could therefore
change depending on "tselect".

Remove this ambiguity by rewording the description of "tinfo" CSR
and "tinfo.version" field.